requirements.txt
pyaes==1.6.1
pyasn1==0.6.0
PySocks==1.7.1
python-dotenv==1.0.1
rsa==4.9
setuptools==69.2.0
Telethon==1.34.0
